dummy-python-service/service/requirements.txt
python-dateutil==2.8.2
pymongo==4.5.0
fastapi==0.103.2
pydantic==2.4.2
gunicorn==21.2.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
python-logging-loki==0.3.1

